refactor ApiKeys adding separate ApiKeysHeader and ApiKeysCouple reusable components
