+++ apps/backoffice/src/components/api-keys/api-keys.tsx
@@ -1,10 +1,9 @@
 import { SubscriptionKeyTypeEnum } from "@/generated/api/SubscriptionKeyType";
 import { SubscriptionKeys } from "@/generated/api/SubscriptionKeys";
-import { Box, Divider, Typography } from "@mui/material";
-import { useTranslation } from "next-i18next";
+import { Box } from "@mui/material";
 
-import { useDialog } from "../dialog-provider";
-import { ApiSingleKey } from "./api-single-key";
+import { ApiKeysCouple } from "./api-keys-couple";
+import { ApiKeysHeader } from "./api-keys-header";
 
 export interface ApiKeysProps {
   /** Main component card description */
@@ -14,59 +13,21 @@ export interface ApiKeysProps {
   /** Event triggered when user click "Rotate" on confirmation modal */
   onRotateKey: (type: SubscriptionKeyTypeEnum) => void;
   /** Main component card title */
-  title: string;
+  title?: string;
 }
 
-/** API Keys main component
+/** API Key main component
  *
- * Used to show, copy, rotate `keys` _(primary/secondary)_ and edit optional `authorized cidrs`.
+ * Used to show, copy, rotate `keys` _(primary/secondary)_.
  * */
-export const ApiKeys = (props: ApiKeysProps) => {
-  const { t } = useTranslation();
-  const showDialog = useDialog();
-
-  const handleRotate = async (keyType: SubscriptionKeyTypeEnum) => {
-    const makeKeyRotation = await showDialog({
-      confirmButtonLabel: t("keys.rotate.button"),
-      message: t("keys.rotate.modal.description"),
-      title: t("keys.rotate.modal.title"),
-    });
-    if (makeKeyRotation) {
-      props.onRotateKey(keyType);
-    } else {
-      console.warn("Operation canceled");
-    }
-  };
-
-  return (
-    <Box
-      bgcolor="background.paper"
-      borderRadius={0.5}
-      id="card-details"
-      padding={3}
-    >
-      <Typography variant="h6">{t(props.title)}</Typography>
-      <Typography
-        color="text.secondary"
-        marginBottom={3}
-        marginTop={1}
-        variant="body2"
-      >
-        {props.description ? t(props.description) : ""}
-      </Typography>
-      <ApiSingleKey
-        keyType={SubscriptionKeyTypeEnum.primary}
-        keyValue={props.keys?.primary_key}
-        onBlockClick={(kt) => console.log("onBlockClick:", kt)}
-        onRotateClick={handleRotate}
-      />
-      <Divider sx={{ marginTop: 3 }} />
-      <ApiSingleKey
-        keyType={SubscriptionKeyTypeEnum.secondary}
-        keyValue={props.keys?.secondary_key}
-        onBlockClick={(kt) => console.log("onBlockClick:", kt)}
-        onRotateClick={handleRotate}
-      />
-    </Box>
-  );
-};
+export const ApiKeys = (props: ApiKeysProps) => (
+  <Box
+    bgcolor="background.paper"
+    borderRadius={0.5}
+    id="card-details"
+    padding={3}
+  >
+    <ApiKeysHeader description={props.description} title={props.title} />
+    <ApiKeysCouple keys={props.keys} onRotateKey={props.onRotateKey} />
+  </Box>
+);
