+++ apps/backoffice/src/components/api-keys/index.ts
@@ -1,4 +1,7 @@
 export * from "./api-key-value";
 export * from "./api-keys";
+export * from "./api-keys-couple";
+export * from "./api-keys-header";
 export * from "./api-single-key";
 export * from "./authorized-cidrs";
+export * from "./groups-api-keys";
